feat: changed location of base class for websocket api
